refactor(changelog-generator): extract printBanner() into separate file
To declutter.
