+++ projects/npm-tools/packages/changelog-generator/src/index.js
@@ -9,6 +9,7 @@ const {promisify} = require('util');
 
 const {cleanup, error, info, log, prompt, warn} = require('./console');
 const git = require('./git');
+const printBanner = require('./printBanner');
 const readYarnrc = require('./readYarnrc');
 
 const readFileAsync = promisify(fs.readFile);
@@ -251,56 +252,6 @@ async function formatChanges(changes, remote) {
 		})
 		.filter(Boolean)
 		.join('\n\n');
-}
-
-/**
- * Prints `message` in a silly banner like this:
- *  ____________________________________
- * (_)                                  `
- *   |                                  |
- *   |   @liferay/changelog-generator   |
- *   |   ============                   |
- *   |                                  |
- *   |   Reporting for duty!            |
- *   |                                  |
- *   |__________________________________|
- *   (_)________________________________)
- *
- */
-function printBanner(message) {
-	const lines = message.split('\n').map((line) => line.trim());
-
-	const width = lines.reduce((max, line) => {
-		return Math.max(max, line.length);
-	}, 0);
-
-	const TEMPLATE = [
-		[' _____', '_', '___  '],
-		['(_)   ', ' ', '   ` '],
-		['  |   ', '*', '   | '],
-		['  |___', '_', '___| '],
-		['  (_)_', '_', '____)'],
-	];
-
-	let banner = '';
-
-	TEMPLATE.forEach(([left, middle, right]) => {
-		if (middle === '*') {
-			lines.forEach((line) => {
-				banner +=
-					left +
-					line +
-					' '.repeat(width - line.length) +
-					right +
-					'\n';
-			});
-		}
-		else {
-			banner += left + middle.repeat(width) + right + '\n';
-		}
-	});
-
-	log(banner);
 }
 
 function relative(filePath) {
